refactor(domain): BoardId를 Long 타입으로 수정

+++ src/main/java/domain/board/BoardSelectCommand.java
@@ -1,15 +1,15 @@
 package domain.board;
 
 public record BoardSelectCommand(
-        int select
+        long select
 ) {
 
     private static final int SELECT_COMMAND_THRESHOLD = 0;
 
     public static BoardSelectCommand from(String input) {
         validateStringIsNumeric(input);
-        validateCommandRange(Integer.parseInt(input));
-        return new BoardSelectCommand(Integer.parseInt(input));
+        validateCommandRange(Long.parseLong(input));
+        return new BoardSelectCommand(Long.parseLong(input));
     }
 
     public static void validateStringIsNumeric(String input) {
@@ -22,7 +22,7 @@ public record BoardSelectCommand(
     }
 
     // TODO 커스텀 예외로 변경할 것.
-    public static void validateCommandRange(int select) {
+    public static void validateCommandRange(long select) {
         if (select < SELECT_COMMAND_THRESHOLD) {
             throw new IllegalArgumentException();
         }
